libnagios: Fix warnings about signed-vs-unsigned comparisons

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2735 5f96b256-904b-4d8d-8c98-d829582c6739

+++ lib/iobroker.c
@@ -84,7 +84,7 @@ const char *iobroker_strerror(int error)
 	if (error < 0) {
 		return iobroker_unknown_error;
 	}
-	if (error >= ARRAY_SIZE(iobroker_errors))
+	if (error >= (int)ARRAY_SIZE(iobroker_errors))
 		return strerror(error);
 
 	return iobroker_errors[error].string;
